requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
psycopg2-binary==2.9.7
redis==5.0.1
bcrypt==4.1.2
cryptography==41.0.7
python-dotenv==1.0.0
marshmallow==3.20.1
email-validator==2.1.0
phonenumbers==8.13.25
pytest==7.4.3
pytest-flask==1.3.0
requests==2.31.0
celery==5.3.4
gunicorn==21.2.0

# Additional packages for monitoring and security
psutil==5.9.6
PyJWT==2.8.0
passlib==1.7.4
argon2-cffi==23.1.0
structlog==23.2.0
sentry-sdk[flask]==1.38.0
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
coverage==4.1.0